refactor(velvet): consolidate Routing, Store, and Preview test fixtures
Merge single-seam fixture files into topical fixtures (6 clusters:
blocker liveness, route matching, relative navigation, routing hooks,
store notifier re-entrancy, preview viewport/zoom). Test bodies are
preserved verbatim; colocation is used where nested helper types or
EditorPrefs-driven setups differ.
