--- Packages/com.velvet.core/Editor/Preview/Tests/Editor/VelvetPreviewWindowZoomLayoutTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d52965a1ea8e544d3827f527c35c2604
