--- Packages/com.velvet.core/Runtime/Routing/Tests/Editor/BlockerLivenessTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 27cbd63b3aaa04ec3b691269d45cf125
