--- Packages/com.velvet.core/Runtime/Routing/Tests/Editor/RouteMatchingTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 274daf63d7dc7473180575bff9054409
