--- Packages/com.velvet.core/Runtime/Routing/Tests/Editor/RouteNavigateHookTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3e6fbf747cf8f4803b34d69e04cd3df8
