--- Packages/com.velvet.core/Runtime/Routing/Tests/Editor/RouteNavigationStateTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8e81657de285d42fe8340e62958dc74d
