--- Packages/com.velvet.core/Runtime/Routing/Tests/Editor/RouterRouteRelativeNavigationTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 503d46171570840dfb76f601d2b74716
