--- Packages/com.velvet.core/Runtime/Store/Tests/Editor/StoreReentrantNotifyTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 468d60c62542c49d290a08fb5e3f09fe
